Refactor network
