+++ lib/kaspa/kaspa.dart
@@ -2,6 +2,7 @@ export 'address_generator.dart';
 export 'client/kaspa_client.dart';
 export 'kaspa_api/kaspa_api.dart';
 export 'kaspa_api/kaspa_api_service.dart';
+export 'network.dart';
 export 'signer_base.dart';
 export 'transaction/types.dart';
 export 'types.dart';
