+++ lib/kaspa/network.dart
@@ -1,6 +1,54 @@
-import 'package:coinslib/coinslib.dart';
+import 'package:coinslib/coinslib.dart' show NetworkType, Bip32Type;
 
-import '../kaspa.dart';
+const int kMainnetRpcPort = 16110;
+const int kTestnetPpcPort = 16210;
+const int kSimnetRpcPort = 16510;
+const int kDevnetRpcPort = 16610;
+
+enum KaspaNetwork {
+  mainnet,
+  testnet,
+  devnet,
+  simnet,
+}
+
+int portForNetwork(KaspaNetwork network) {
+  switch (network) {
+    case KaspaNetwork.mainnet:
+      return kMainnetRpcPort;
+    case KaspaNetwork.testnet:
+      return kTestnetPpcPort;
+    case KaspaNetwork.simnet:
+      return kSimnetRpcPort;
+    case KaspaNetwork.devnet:
+      return kDevnetRpcPort;
+  }
+}
+
+KaspaNetwork networkForPort(int port) {
+  switch (port) {
+    case kMainnetRpcPort:
+      return KaspaNetwork.mainnet;
+    case kTestnetPpcPort:
+      return KaspaNetwork.testnet;
+    case kSimnetRpcPort:
+      return KaspaNetwork.simnet;
+    case kDevnetRpcPort:
+      return KaspaNetwork.devnet;
+    default:
+      return KaspaNetwork.mainnet;
+  }
+}
+
+KaspaNetwork networkForKpub(String kpub) {
+  return switch (kpub.substring(0, 4)) {
+    'kpub' => KaspaNetwork.mainnet,
+    'ktub' => KaspaNetwork.testnet,
+    'ksub' => KaspaNetwork.simnet,
+    'kdub' => KaspaNetwork.devnet,
+    _ => KaspaNetwork.mainnet,
+  };
+}
 
 NetworkType networkTypeForNetwork(KaspaNetwork network) {
   switch (network) {
