+++ lib/kaspa/types.dart
@@ -3,6 +3,5 @@ export 'types/address.dart';
 export 'types/address_balance.dart';
 export 'types/address_prefix.dart';
 export 'types/amount.dart';
-export 'types/network.dart';
-export 'types/token_info.dart';
 export 'types/kaspa_uri.dart';
+export 'types/token_info.dart';
