--- lib/kaspa/wallet/keychains.dart
@@ -1,2 +1,0 @@
-const kExternalKeychain = 0;
-const kInternalKeychain = 1;
